core/portal/requirements.txt
flask
requests
python-jose[cryptography]
python-multipart
psycopg2-binary
sqlalchemy
flask-sqlalchemy
